Adding the test-callbacl test.
