+++ test/index.js
@@ -25,25 +25,4 @@ describe('Blankie', () => {
             }
         }])).to.reject(Error, 'child "reportOnly" fails because ["reportOnly" must be a boolean]');
     });
-
-    it.only('allows a callback as the only option', async () => {
-
-        const cspCallback = function () {
-
-            const options = {};
-            options['base-uri'] = 'self';
-            return options;
-        };
-        const server = Hapi.server();
-        await server.register([Scooter, {
-            plugin: Blankie,
-            options: cspCallback
-        }]);
-        const res = await server.inject({
-            method: 'GET',
-            url: '/'
-        });
-        expect(res.headers).to.contain('content-security-policy');
-        expect(res.headers['content-security-policy']).to.contain('base-uri \'self');
-    });
 });
